updog: use u32 seed to match new definition

+++ workspaces/updater/updog/src/de.rs
@@ -12,17 +12,17 @@ use std::str::FromStr;
 /// Converts the bound key to an integer before insertion and catches duplicates
 pub(crate) fn deserialize_bound<'de, D>(
     deserializer: D,
-) -> Result<BTreeMap<u64, DateTime<Utc>>, D::Error>
+) -> Result<BTreeMap<u32, DateTime<Utc>>, D::Error>
 where
     D: Deserializer<'de>,
 {
     fn bound_to_int(
         key: String,
         time: DateTime<Utc>,
-        map: &mut BTreeMap<u64, DateTime<Utc>>,
+        map: &mut BTreeMap<u32, DateTime<Utc>>,
     ) -> Result<(), error::Error> {
         let bound = key
-            .parse::<u64>()
+            .parse::<u32>()
             .context(error::BadBound { bound_str: key })?;
         ensure!(
             map.insert(bound, time).is_none(),
@@ -35,7 +35,7 @@ where
     struct Visitor;
 
     impl<'de> serde::de::Visitor<'de> for Visitor {
-        type Value = BTreeMap<u64, DateTime<Utc>>;
+        type Value = BTreeMap<u32, DateTime<Utc>>;
 
         fn expecting(&self, formatter: &mut fmt::Formatter<'_>) -> fmt::Result {
             formatter.write_str("a map")
